Moved some more expression simplification code to hwutils
